+++ tests/Fixtures/ImportedFromGoogle.php
@@ -8,22 +8,32 @@
 
 namespace Piwik\Plugins\GoogleAnalyticsImporter\tests\Fixtures;
 
+use Piwik\Archive\ArchiveInvalidator;
 use Piwik\Config;
+use Piwik\Container\StaticContainer;
 use Piwik\CronArchive;
+use Piwik\DataAccess\ArchiveTableCreator;
 use Piwik\Db;
+use Piwik\DbHelper;
 use Piwik\Ini\IniReader;
 use Piwik\Option;
+use Piwik\Date;
 use Piwik\Plugins\GoogleAnalyticsImporter\Google\Authorization;
+use Piwik\Plugins\GoogleAnalyticsImporter\ImportStatus;
 use Piwik\Plugins\GoogleAnalyticsImporter\tests\Framework\CapturingGoogleClient;
+use Piwik\Plugins\GoogleAnalyticsImporter\tests\Framework\MockResponseClient;
 use Piwik\Plugins\VisitsSummary\API;
+use Piwik\Plugins\VisitsSummary\VisitsSummary;
 use Piwik\Tests\Framework\Fixture;
+use Piwik\Timer;
 use Symfony\Bridge\Monolog\Handler\ConsoleHandler;
 
 class ImportedFromGoogle extends Fixture
 {
     public $idSite = 1;
     public $dateTime = '2019-06-30 00:00:00';
-    public $importedDateRange = '2019-06-27,2019-07-02';
+    public $importedDateRange1 = '2019-06-27,2019-07-01';
+    public $importedDateRange2 = '2019-07-02,2019-07-02';
 
     public $campaignIdSite = 2;
     public $campaignDataDateTime = '2019-08-14';
@@ -47,7 +57,13 @@ class ImportedFromGoogle extends Fixture
 
         $this->getGoogleAnalyticsParams();
 
-        $this->runGoogleImporter($this->importedDateRange);
+        $this->runGoogleImporter($this->importedDateRange1);
+        $this->extendEndDate($idSite = 1, '2019-07-02');
+        $this->scheduleReimport($idSite = 1, '2019-06-27', '2019-06-27');
+
+        $output = $this->runGoogleImporter($this->importedDateRange2, $idSite = 1);
+        $this->assertContains('Importing reports for date range 2019-06-27 - 2019-06-27 from GA view', $output);
+
         $this->runGoogleImporter($this->campaignDataDateRange);
 
         $this->aggregateForYear();
@@ -55,30 +71,47 @@ class ImportedFromGoogle extends Fixture
         // track a visit on 2019-07-03 and make sure it appears correctly in reports
         $this->trackVisitAfterImport();
 
-        $this->invalidateArchives();
+        $this->invalidateArchives(); // trigger core:archive invalidation
         $this->aggregateForYear();
+
+        $this->invalidateAndRearchiveDay(); // make sure day archives won't be re-archived
+
+        print "Done aggregating.\n";
     }
 
     private function invalidateArchives()
     {
+        ArchiveTableCreator::$tablesAlreadyInstalled = null;
+        DbHelper::getTablesInstalled(true);
+
         $cronArchive = new CronArchive();
         $cronArchive->invalidateArchivedReportsForSitesThatNeedToBeArchivedAgain();
     }
 
     private function aggregateForYear()
     {
-        API::getInstance()->get($this->idSite, 'year', '2019-06-27');
+        $original = @$_GET['trigger'];
+        $_GET['trigger'] = 'archivephp';
+        try {
+            API::getInstance()->get($this->idSite, 'year', '2019-06-27');
+        } finally {
+            if (!empty($original)) {
+                $_GET['trigger'] = $original;
+            } else {
+                unset($_GET['trigger']);
+            }
+        }
     }
 
     private function trackVisitAfterImport()
     {
         $t = Fixture::getTracker($this->idSite, '2019-07-03 05:23:45');
         $t->setUrl('http://matthieu.net/normal/visit/');
-        $t->doTrackPageView('normal visit');
+        Fixture::checkResponse($t->doTrackPageView('normal visit'));
 
         $t->setForceVisitDateTime('2019-07-03 05:25:45');
         $t->setUrl('http://matthieu.net/blog/inde-par-region-et-ville/');
-        $t->doTrackPageView('inde par region et ville');
+        Fixture::checkResponse($t->doTrackPageView('inde par region et ville'));
     }
 
     private function getGoogleAnalyticsParams()
@@ -107,7 +140,7 @@ class ImportedFromGoogle extends Fixture
         return $value;
     }
 
-    private function runGoogleImporter($dates)
+    private function runGoogleImporter($dates, $idSiteToResume = null)
     {
         $domain = Config::getHostname();
         $domainParam = $domain ? ('--matomo-domain=' . $domain) : '';
@@ -117,8 +150,15 @@ class ImportedFromGoogle extends Fixture
         Option::set(Authorization::CLIENT_CONFIG_OPTION_NAME, $this->clientConfig);
 
         $command = "php " . PIWIK_INCLUDE_PATH . '/tests/PHPUnit/proxy/console ' . $domainParam
-            . ' googleanalyticsimporter:import-reports --view=' . $this->viewId
-            . ' --dates=' . $dates . ' --property=' . $property . ' --extra-custom-dimension=ga:networkLocation,visit';
+            . ' googleanalyticsimporter:import-reports ';
+        if ($idSiteToResume) {
+            $command .= '--idsite=' . $idSiteToResume;
+        } else {
+            $command .= '--view=' . $this->viewId
+                . ' --dates=' . $dates . ' --property=' . $property . ' --extra-custom-dimension=ga:networkLocation,visit';
+        }
+
+        $timer = new Timer();
 
         print "\nImporting from google...\n";
 
@@ -132,6 +172,10 @@ class ImportedFromGoogle extends Fixture
         if (strpos($allOutput, 'Encountered unknown') !== false) {
             throw new \Exception("Found problem warning in GA Import output: " . $allOutput);
         }
+
+        print "Done in $timer\n";
+
+        return $allOutput;
      }
 
     public function provideContainerConfig()
@@ -142,6 +186,11 @@ class ImportedFromGoogle extends Fixture
                 \DI\get(ConsoleHandler::class),
             ],
         ];
+
+        if (getenv('MATOMO_USE_MOCK_RESPONSE')) {
+            MockResponseClient::$isForSystemTest = true;
+            $result['GoogleAnalyticsImporter.googleClientClass'] = MockResponseClient::class;
+        }
 
         if ($this->isCapturingResponses) {
             $result['GoogleAnalyticsImporter.googleClientClass'] = CapturingGoogleClient::class;
@@ -180,6 +229,42 @@ class ImportedFromGoogle extends Fixture
         } finally {
             Db::destroyDatabaseObject();
             Config::getInstance()->database = $dbConfig;
+        }
+    }
+
+    private function scheduleReimport($idSite, $start, $end)
+    {
+        /** @var ImportStatus $importStatus */
+        $importStatus = StaticContainer::get(ImportStatus::class);
+        $importStatus->reImportDateRange($idSite, Date::factory($start), Date::factory($end));
+    }
+
+    private function extendEndDate($idSite, $endDate)
+    {
+        /** @var ImportStatus $importStatus */
+        $importStatus = StaticContainer::get(ImportStatus::class);
+        $importStatus->setImportDateRange($idSite, null, Date::factory($endDate));
+    }
+
+    private function invalidateAndRearchiveDay()
+    {
+        ArchiveTableCreator::$tablesAlreadyInstalled = null;
+        DbHelper::getTablesInstalled(true);
+
+        /** @var ArchiveInvalidator $archiveInvalidator */
+        $archiveInvalidator = StaticContainer::get(ArchiveInvalidator::class);
+        $archiveInvalidator->markArchivesAsInvalidated([$this->idSite], [Date::factory('2019-06-28')], 'day');
+
+        $original = @$_GET['trigger'];
+        $_GET['trigger'] = 'archivephp';
+        try {
+            API::getInstance()->get($this->idSite, 'day', '2019-06-28');
+        } finally {
+            if (!empty($original)) {
+                $_GET['trigger'] = $original;
+            } else {
+                unset($_GET['trigger']);
+            }
         }
     }
 }